FIX: Admins unable to cancel a subscription for a user
Stripe at some point deprecated their `.delete` method on Subscription
in favor of `.cancel`. When we upgraded to latest stripe gem in dcde03d7c46661e447c22524ab376c95698b7517
this call likely stopped working.

In addition to this fix we are no longer deleting the customer record,
but instead of updating the status of it. This way an admin can easily
see 'active' as well as 'canceled' subscriptions. This behavior matches
what we are already doing for an individual user listing all of their
subscriptions.

Bug report: https://meta.discourse.org/t/316007